Update emotions.py
Changed implementation of emotion detection for higher accuracy. Now keeps track of an emotion array and only exit with emotion code when three consecutive emotions are the same. Should not affect integration because it still exits with emotion code. No longer exit with -1 though, since it doesn't exit until accurate and consistent emotion is detected.